MeetingScheduler : restructured few classes and added a readme.md file which will have requirements details
