+++ LLDandDesingPatterns/src/T_MeetingScheduler/NotificationDetails/NotificationDetails.java
@@ -1,4 +1,4 @@
-package T_MeetingScheduler;
+package T_MeetingScheduler.NotificationDetails;
 
 public class NotificationDetails {
