+++ LLDandDesingPatterns/src/T_MeetingScheduler/NotificationStrategy/EmailNotification.java
@@ -1,6 +1,6 @@
 package T_MeetingScheduler.NotificationStrategy;
 
-import T_MeetingScheduler.NotificationDetails;
+import T_MeetingScheduler.NotificationDetails.NotificationDetails;
 
 public class EmailNotification implements NotificationStrategy {
